# Add mask parameter to find_contours