--- a/skimage/measure/_find_contours.py
+++ b/skimage/measure/_find_contours.py
@@ -8,7 +8,7 @@ _param_options = ('high', 'low')
 
 def find_contours(array, level,
                   fully_connected='low', positive_orientation='low',
-                  nodata=None):
+                  mask=None):
     """Find iso-valued contours in a 2D array for a given level value.
 
     Uses the "marching squares" method to compute a the iso-valued contours of
@@ -31,11 +31,10 @@ def find_contours(array, level,
          contours will wind counter- clockwise around elements below the
          iso-value. Alternately, this means that low-valued elements are always
          on the left of the contour. (See below for details.)
-    nodata : float
-        Value to treat as missing data. No contours will be drawn where *array*
-        has values equal to *nodata* (or where values are ``NaN`` if nodata is
-        ``NaN``).  Default value is ``None``, which disables this behaviour
-        entirely for backwards-compatibility.
+    mask : 2D ndarray of bool, or None
+        A boolean mask, True where we want to draw contours.
+        Note that NaN values are always excluded from the considered region (
+        i.e. mask=False is forced where where array is NaN).
 
     Returns
     -------
@@ -64,9 +63,10 @@ def find_contours(array, level,
     with the 'fully_connected' parameter.
 
     Output contours are not guaranteed to be closed: contours which intersect
-    the array edge or a region of missing data will be left open. All other
-    contours will be closed. (The closed-ness of a contours can be tested by
-    checking whether the beginning point is the same as the end point.)
+    the array edge or a masked-off region (either where mask is False or where
+    array is NaN) will be left open. All other contours will be closed. (The
+    closed-ness of a contours can be tested by checking whether the beginning
+    point is the same as the end point.)
 
     Contours are oriented. By default, array values lower than the contour
     value are to the left of the contour and values greater than the contour
@@ -117,15 +117,19 @@ def find_contours(array, level,
     if array.ndim != 2:
         raise ValueError('Only 2D arrays are supported.')
     level = float(level)
-    if nodata is not None:
-        nodata = np.double(nodata)
+    if mask is None:
+        mask = np.ones_like(array, dtype=bool)
+    else:
+        mask = np.asarray(mask, dtype=bool)
+        if mask.shape != array.shape:
+            raise ValueError('Parameters "array" and "mask" must have same shape.')
     if (fully_connected not in _param_options or
        positive_orientation not in _param_options):
         raise ValueError('Parameters "fully_connected" and'
         ' "positive_orientation" must be either "high" or "low".')
     point_list = _find_contours_cy.iterate_and_store(array, level,
                                                      fully_connected == 'high',
-                                                     nodata=nodata)
+                                                     mask=mask)
     contours = _assemble_contours(_take_2(point_list))
     if positive_orientation == 'high':
         contours = [c[::-1] for c in contours]