--- a/skimage/measure/_find_contours.py
+++ b/skimage/measure/_find_contours.py
@@ -117,9 +117,7 @@ def find_contours(array, level,
     if array.ndim != 2:
         raise ValueError('Only 2D arrays are supported.')
     level = float(level)
-    if mask is None:
-        mask = np.ones_like(array, dtype=np.uint8)
-    else:
+    if mask is not None:
         mask = np.asarray(mask, dtype=np.uint8)
         if mask.shape != array.shape:
             raise ValueError('Parameters "array" and "mask" must have same shape.')