--- a/skimage/measure/_find_contours.py
+++ b/skimage/measure/_find_contours.py
@@ -118,6 +118,8 @@ def find_contours(array, level,
         raise ValueError('Only 2D arrays are supported.')
     level = float(level)
     if mask is not None:
+        if not np.can_cast(mask.dtype, bool, casting='safe'):
+            raise TypeError('Parameter "mask" must be a binary array.')
         mask = np.asarray(mask, dtype=np.uint8)
         if mask.shape != array.shape:
             raise ValueError('Parameters "array" and "mask" must have same shape.')