--- a/skimage/measure/_find_contours.py
+++ b/skimage/measure/_find_contours.py
@@ -122,7 +122,8 @@ def find_contours(array, level,
             raise TypeError('Parameter "mask" must be a binary array.')
         mask = np.asarray(mask, dtype=np.uint8)
         if mask.shape != array.shape:
-            raise ValueError('Parameters "array" and "mask" must have same shape.')
+            raise ValueError('Parameters "array" and "mask"'
+                             ' must have same shape.')
     if (fully_connected not in _param_options or
        positive_orientation not in _param_options):
         raise ValueError('Parameters "fully_connected" and'