--- a/skimage/measure/tests/test_find_contours.py
+++ b/skimage/measure/tests/test_find_contours.py
@@ -9,6 +9,9 @@ a = np.ones((8, 8), dtype=np.float32)
 a[1:-1, 1] = 0
 a[1, 1:-1] = 0
 
+b = np.copy(a)
+b[7:,:] = np.nan
+
 x, y = np.mgrid[-1:1:5j, -1:1:5j]
 r = np.sqrt(x**2 + y**2)
 
@@ -44,6 +47,33 @@ def test_binary():
     assert len(contours) == 1
     assert_array_equal(contours[0][::-1], ref)
 
+def test_nodata():
+    ref = [[6. ,  1.5],
+           [5. ,  1.5],
+           [4. ,  1.5],
+           [3. ,  1.5],
+           [2. ,  1.5],
+           [1.5,  2. ],
+           [1.5,  3. ],
+           [1.5,  4. ],
+           [1.5,  5. ],
+           [1.5,  6. ],
+           [1. ,  6.5],
+           [0.5,  6. ],
+           [0.5,  5. ],
+           [0.5,  4. ],
+           [0.5,  3. ],
+           [0.5,  2. ],
+           [0.5,  1. ],
+           [1. ,  0.5],
+           [2. ,  0.5],
+           [3. ,  0.5],
+           [4. ,  0.5],
+           [5. ,  0.5],
+           [6. ,  0.5]]
+    contours = find_contours(b, 0.5, positive_orientation='high', nodata=np.nan)
+    assert len(contours) == 1
+    assert_array_equal(contours[0][::-1], ref)
 
 def test_float():
     contours = find_contours(r, 0.5)