--- a/skimage/measure/tests/test_find_contours.py
+++ b/skimage/measure/tests/test_find_contours.py
@@ -9,9 +9,6 @@ a = np.ones((8, 8), dtype=np.float32)
 a[1:-1, 1] = 0
 a[1, 1:-1] = 0
 
-b = np.copy(a)
-b[7:, :] = np.nan
-
 x, y = np.mgrid[-1:1:5j, -1:1:5j]
 r = np.sqrt(x**2 + y**2)
 
@@ -48,34 +45,56 @@ def test_binary():
     assert_array_equal(contours[0][::-1], ref)
 
 
+# target contour for mask tests
+mask_contour = [
+    [6. ,  0.5],
+    [5. ,  0.5],
+    [4. ,  0.5],
+    [3. ,  0.5],
+    [2. ,  0.5],
+    [1. ,  0.5],
+    [0.5,  1. ],
+    [0.5,  2. ],
+    [0.5,  3. ],
+    [0.5,  4. ],
+    [0.5,  5. ],
+    [0.5,  6. ],
+    [1. ,  6.5],
+    [1.5,  6. ],
+    [1.5,  5. ],
+    [1.5,  4. ],
+    [1.5,  3. ],
+    [1.5,  2. ],
+    [2. ,  1.5],
+    [3. ,  1.5],
+    [4. ,  1.5],
+    [5. ,  1.5],
+    [6. ,  1.5],
+]
+
+mask = np.ones((8, 8), dtype=bool)
+# Some missing data that should result in target hole in the contour:
+mask[7, 0:3] = False
+
+# Some missing data that shouldn't change anything:
+mask[0, 7] = False
+mask[2, 7] = False
+
+
 def test_nodata():
-    ref = [[6. ,  1.5],
-           [5. ,  1.5],
-           [4. ,  1.5],
-           [3. ,  1.5],
-           [2. ,  1.5],
-           [1.5,  2. ],
-           [1.5,  3. ],
-           [1.5,  4. ],
-           [1.5,  5. ],
-           [1.5,  6. ],
-           [1. ,  6.5],
-           [0.5,  6. ],
-           [0.5,  5. ],
-           [0.5,  4. ],
-           [0.5,  3. ],
-           [0.5,  2. ],
-           [0.5,  1. ],
-           [1. ,  0.5],
-           [2. ,  0.5],
-           [3. ,  0.5],
-           [4. ,  0.5],
-           [5. ,  0.5],
-           [6. ,  0.5]]
-    contours = find_contours(b, 0.5, positive_orientation='high',
-                             nodata=np.nan)
+    # Test missing data via NaNs in input array
+    b = np.copy(a)
+    b[~mask] = np.nan
+    contours = find_contours(b, 0.5, positive_orientation='high')
     assert len(contours) == 1
-    assert_array_equal(contours[0][::-1], ref)
+    assert_array_equal(contours[0], mask_contour)
+
+
+def test_mask():
+    # Test missing data via explicit masking
+    contours = find_contours(a, 0.5, positive_orientation='high', mask=mask)
+    assert len(contours) == 1
+    assert_array_equal(contours[0], mask_contour)
 
 
 def test_float():