## Description

Resolves #3006: Adds a `nodata` parameter to `measure.find_contours` which draws no contours in the region where the input array is equal to `nodata`.

`find_contours` currently misbehaves when handling input arrays containing `NaN` values: it attempts to linearly interpolate these values and ends up returning contours containing `NaN` coordinates. This PR allows avoiding this by passing `nodata=numpy.nan`. For now I've left the previous behaviour as the default (`nodata=None`) for backwards compatibility, in case anyone was relying on it.

The implementation is very simple - no lines are added in any squares with at least one `NaN` corner. This could be cleverer - for example, the square `[[1,NaN],[0,1]]` should probably result in the same segment as `[[1,1],[0,1]]` - but I didn't have time to think through all the possible combinations.

## Checklist

<!-- It's fine to submit PRs which are a work in progress! -->
<!-- But before they are merged, all PRs should provide: -->
- [x] [Docstrings for all functions](https://github.com/numpy/numpy/blob/master/doc/example.py)
- [ ] Gallery example in `./doc/examples` (new features only)
- [ ] Benchmark in `./benchmarks`, if your changes aren't covered by an
  existing benchmark
- [x] Unit tests
- [ ] Clean style in [the spirit of PEP8](https://www.python.org/dev/peps/pep-0008/)

<!-- For detailed information on these and other aspects see -->
<!-- the scikit-image contribution guidelines. -->
<!-- https://scikit-image.org/docs/dev/contribute.html -->

## For reviewers

<!-- Don't remove the checklist below. -->
- [ ] Check that the PR title is short, concise, and will make sense 1 year
  later.
- [ ] Check that new functions are imported in corresponding `__init__.py`.
- [ ] Check that new features, API changes, and deprecations are mentioned in
      `doc/release/release_dev.rst`.
- [ ] Consider backporting the PR with `@meeseeksdev backport to v0.14.x`