@acarapetis 2019-09-19
Ran a quick and dirty benchmark to see the performance impact of the nan checks when dealing with nan-free data:
```python
# Once, to generate fixed test data:
for s in [20,50,100,500,1000,2000,3000]:
    np.save(f'{s}.npy', np.random.random((s,s)))

# Then once for each version of the code:
import numpy as np
data = [(s, np.load(f'{s}.npy')) for s in [20,50,100,500,1000,2000,3000]]
from skimage.measure import find_contours
for s, arr in data:
    print(f'{s}x{s} ', end='')
    %timeit find_contours(np.random.random((s,s)), 0.5)
```

Not super controlled conditions, but on my hardware at least it looks like roughly a 2% hit:

| size | before | after | difference |
| --- | --- | --- | --- |
| 20x20 | 714 µs ± 1.87 µs | 727 µs ± 6.15 µs | ~2% slower |
| 50x50 | 4.47 ms ± 23.2 µs | 4.52 ms ± 10.5 µs | ~1% slower |
| 100x100 | 19.2 ms ± 244 µs | 19.2 ms ± 299 µs | ~no change |
| 500x500 | 539 ms ± 3.91 ms | 549 ms ± 10.5 ms | ~2% slower |
| 1000x1000 | 2.23 s ± 15.2 ms | 2.27 s ± 27.3 ms | ~2% slower |
| 2000x2000 | 9.38 s ± 194 ms | 9.35 s ± 51.8 ms | ~0.3% faster |
| 3000x3000 | 21.7 s ± 328 ms | 22.2 s ± 965 ms | ~2% slower |